Modularization:
ScreenEngine manages the menus.
GAME.screens['main-menu'] in screens_main.js contains the major
interface between the GAME and ScreenEngine.

Now, the ScreenEngine can be called to end the game or whatever.
That is, when GAME makes use of ScreenEngine.

NOTE: now ScreenEngine is invoked from Loader.js when all preloading
is comlplete

Signed-off-by: git-us3r <mlink3.14@gmail.com>